Add a parameter that allows to skip data conversion of a table row if a condition evaluates to true
