+++ phpmd.xml.dist
@@ -27,6 +27,10 @@
     <rule ref="rulesets/codesize.xml"/>
     <rule ref="rulesets/controversial.xml"/>
     <rule ref="rulesets/design.xml"/>
-    <rule ref="rulesets/naming.xml"/>
+
+    <rule ref="rulesets/naming.xml">
+        <exclude name="LongVariable"/>
+    </rule>
+
     <rule ref="rulesets/unusedcode.xml"/>
 </ruleset>
