+++ src/Dumper/Mysqldump/ExtensionInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Smile\GdprDump\Dumper\Mysqldump;
+
+use Ifsnop\Mysqldump\Mysqldump;
+
+interface ExtensionInterface
+{
+    /**
+     * Register the extension.
+     *
+     * @param Mysqldump $dumper
+     */
+    public function register(Mysqldump $dumper);
+}
